1127_1800

+++ Backend/Backend/devices/models.py
@@ -13,7 +13,7 @@ class device(models.Model):
     )
     devicestatus = models.CharField(max_length=20, choices=DEVICE_STATUS_CHOICES, default='active', null=True, blank=True)
     parameter  = models.JSONField(null=True,blank=True)	# EX:LED的{'RGB':[255,255,255]}
-    devicemode = models.IntegerField(default = 1, null = True, blank = True)
+    devicemode = models.FloatField(default = 1, null = True, blank = True)
     openTime = models.DateTimeField(null = True, blank = True)	# 設置開啟時間
     closeTime = models.DateTimeField(null = True, blank = True)	# 設置關閉時間
     boxid = models.ForeignKey(box, on_delete=models.CASCADE, related_name = 'inwhichbox')  # 外鍵關聯到 box，使用 CASCADE 進行相關刪除操作
